requirements.txt
PyPDF2
pandas
openpyxl
pdfplumber
flask
gunicorn
werkzeug
python-dotenv
